Move and merge features tests into proper location
